Show/Hide Elements Based on Session

+++ list.php
@@ -1,12 +1,16 @@
+<?php session_start(); ?>
 <?php include "inc/html-top.php"; ?>
 
   <body class="container-fluid">
       <header class="full-width">
         <h1><a href = "index.php">CSC 174</a></h1>
 <!-- if logged in -->
+<?php if(isset($_SESSION["loggedin"]) && $_SESSION["loggedin"] === true) { ?>
         <a class="sign" href="logout.php">Log Out</a>
 <!-- if NOT logged in -->
+<?php } else { ?>
         <a class="sign" href="login.php">Log In</a>
+<?php } ?>
       </header>
       <?php
       // connect to the database
@@ -34,11 +38,13 @@
               <a href="<?php echo $row['link']; ?>" target="_blank" class="btn btn-primary">Read More</a>
 
 <!-- if logged in -->
+<?php if(isset($_SESSION["loggedin"]) && $_SESSION["loggedin"] === true) { ?>
             <div class = "dropdown">
               <div id = "<?php echo$row['id']; ?>">
                 <a href="edit.php?id=<?php echo $row['id']; ?>">Edit</a>
                 <a onclick="return confirm('Are you sure you want to delete: <?php echo $row["firstname"] . " " . $row["lastname"]; ?>?')" href="delete.php?id=<?php echo $row['id']; ?>">Delete</a>
             </div>
+<?php } ?>
 
 
             </div>
